requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.18
aiofiles==23.2.1
python-dotenv==1.0.1
mercadopago==2.2.2
anthropic==0.84.0
playwright==1.58.0
email-validator==2.2.0
pytest==8.3.0
pytest-asyncio==0.23.0
httpx==0.27.0
